+++ tests/test_extensions/test_blocks/test_details.py
@@ -36,7 +36,7 @@ class TestBlocksDetails(util.MdCase):
             R'''
             /// details
                 type: note
-                $: .other
+                attrs: {class: other}
 
             Some *content*
             ///